+++ src/main/kotlin/br/upf/sistemacompradekeys/dtos/UsuarioResponseDTO.kt
@@ -1,7 +1,7 @@
 package br.upf.sistemaeventos.dtos
 
 data class UsuarioResponseDTO(
-    val cpf: Int,
+    val cpf: Long,
     val username: String,
     val regiao: String,
     val telefone: String,
